Allow for multisig actions
Implement multisig confirmations on any action that is added to the
group. This allows for members to confirm or revoke multiple actions.
In order to conform to the IForwarder interface that the contract
implements we will have to have some helper functions in the script that
allows us to looks up which scripts (actions) the group can take and
then check if that action has enough confirmations.

It may be helpful to add some functions to return which members
confirmed which actions but I figured that could be added later as an
enhancement on top of the current functionality.