Fix addAction function
The addAction function was not incrementing the actionCount varaible
which is used to keep track of how many actions have been added to the
group. This was apparently due to the function trying to return the
actionCount which is the index of the action for easier
reference/lookup. Removing the return if the actionCount allows the
actionCount to properly increment now and we can use the getActionId
function to look up the index for confirmation or revocation of the
action.

Updated the tests to test for the adding of actions and that their
respective indicies are set properly. Also the tests will have to use
the getActionId function to know the proper index for approval.