templates: Fix Finance period bug in Beta Base
Period duration was being set to max uint64, but as this is added to
start time to calculate end time, it was overflowing. Now we substract
current time (which will be start time for the unic existing period)
so the final end period will be max uint 64, achieving the desired
effect of having an endless period.