Voting: avoid re-rendering all votes on each useNow() tick
Unfortunately, `openedStatess is recreated each time, so in effect this
makes the useMemo useless. The string key was kind of a hack to get a
consistent value to compare against.

You can really notice the performance drag this change creates; the entire
app is very sluggish because it's rerendering everything each second
(due to the useNow()).